refactor: Change of folder structure
A bunch of changes to the folder structure within VS. I didn't
like that the solution was called Sandbox, so I renamed it to
AcceleratedC++Exercises to be more reflective of what it actually
is. I also removed the Sandbox project since that wasn't
applicable to the contents of this repo. I also renamed
the Exercises_Section0 to just Chapter0 so that it's now using a
simpler name. No change to production code.